Create README.md
Obsolete vector network analysers (VNAs) in principle allow one to create user’s calibration kits, but only as analytical models including capacitances, inductances, resistances, and delay lines for coaxial terminations. Determing these parameters is a complex task. In modern VNAs, a more convenient option is available when instead of the models for each calibration standard its Touchstone s1p (for SHORT, OPEN, LOAD) or s2p (for THRU) file can be used. The implementation of the SOLT calibration with the s1p/s2p files for the coaxial terminations developed in the present project (see the full report in PDF file) was intended to expand the functionality of VNA HP8753E (30 kHz – 6 GHz) in a research lab at MISiS (https://en.misis.ru/). Experimental setup for measuring high-frequency magneto-impedance in ferromagnetic wires was designed and deployed for this lab by DYK in 2017-2018.